# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/submit/delete_submit": {
            "get": {
                "description": "Delete submit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "submit"
                ],
                "summary": "Delete submit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "delete submit",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
...
        "/user/submit/create_submit": {
            "post": {
                "description": "Create submit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "submit"
                ],
                "summary": "Create submit",
                "parameters": [
                    {
                        "description": "create submit",
                        "name": "createRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Submit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
        "/user/submit/get_submits": {
            "get": {
                "description": "get submit list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "submit"
                ],
                "summary": "get submit list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "page size",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
        "/user/submit/search_submits": {
            "get": {
                "description": "Search submit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "submit"
                ],
                "summary": "Search submit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search submit",
                        "name": "keyword",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "middlewares.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "request.Submit": {
            "type": "object",
            "required": [
                "code",
                "language",
                "question_id"
            ],
            "properties": {
                "code": {
                    "description": "\"用户代码\"",
                    "type": "string",
                    "maxLength": 1024
                },
                "language": {
                    "description": "\"编程语言\"",
                    "type": "string",
                    "maxLength": -9223372036854775808
                },
                "question_id": {
                    "description": "\"判题id\"",
                    "type": "integer",
                    "maximum": 64
                }
            }
        },
...
    }
}